requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
dnspython==2.4.2
email-validator==2.0.0.post2
Faker==19.6.2
fastapi==0.103.2
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
orjson==3.9.7
passlib==1.7.4
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pytest==7.2.1
pytest-asyncio==0.20.3
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
sqladmin==0.15.1
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
WTForms==3.0.1
